--- src/OpenQASM/Untitled-1.jsonc
@@ -1,38 +1,0 @@
-{
-    "0": {
-        "gates": ["SPS","NONE","NONE","CXT",],
-        "id": "cin[0]"
-    },
-    "1": {
-        "gates": ["SPS","X", "CXC","CXC"],
-        "id": "a[0]"
-    },
-    "2": {
-        "gates": ["SPS","X","NONE",],
-        "id": "a[1]"
-    },
-    "3": {
-        "gates": ["SPS","X","NONE",],
-        "id": "a[2]"
-    },
-    "4": {
-        "gates": ["SPS","X","NONE",],
-        "id": "a[3]"
-    },
-    "5": {
-        "gates": ["SPS","X","NONE",],
-        "id": "b[0]"
-    },
-    "6": {
-        "gates": ["SPS","X","NONE",],
-        "id": "b[1]"
-    },
-    "7": {
-        "gates": ["SPS","X","NONE",],
-        "id": "b[2]"
-    },
-    "8": {
-        "gates": ["SPS","X","NONE",],
-        "id": "b[3]"
-    }
-}
